JSCBC-1319: Add exists method to LookupInResult
Motivation
==========
Allow the Node.js SDK to more closely adhere to the RFC.  While the
historical use for the exists() method might not directly apply to the
Node.js client, it is possible some users can find the logic of the
exists() method useful.  Also helps to reduce branching logic w/in FIT
testing (Node is the only SDK that does not implement the exists()
method).

Changes
=======
* Add exists() method to the LookupInResult
* Add test to confirm functionality

Change-Id: I56f6f89a2d24a9620e3f713cf15ee6c407bb759d
Reviewed-on: https://review.couchbase.org/c/couchnode/+/221960
Reviewed-by: Mateusz <matt.wozakowski@couchbase.com>
Tested-by: Jared Casey <jared.casey@couchbase.com>