JSCBC-1363: Split orphan reporter config from tracing config
Motivation
==========
With a future release of the C++ core, the orphan reporting config will
split out from the tracing config.  As the Node client only recently
added the tracing and metrics config it would be ideal to premptively
split the config options in order to prevent having to deprecate the
orphan reporting options w/in the tracing options.

Changes
=======
* Add OrphanReporterConfig
* Update bindings to handle separate orphan reporting config

Change-Id: Ibcb5c574b76fde95815d72f5d33ce0a02ca5b87d
Reviewed-on: https://review.couchbase.org/c/couchnode/+/234372
Tested-by: Jared Casey <jared.casey@couchbase.com>
Reviewed-by: Dimitris Christodoulou <dimitris.christodoulou@couchbase.com>