JSCBC-1367: createBucket and updateBucket are sending unspecified fields to server
Motivation
----------
FIT testing found that we are setting client-side defaults on createBucket and updateBucket
on the minimumDurabilityLevel and BucketType. These should be left up to the server instead.

Changes
-------
minDurabilityLevel set to undefined if not specified by user. Defaults bucket type to 'unknown',
which the C++ core does not encode into the request payload.

Change-Id: I089c520688b28e4ffd27c5f2d40170251846136b
Reviewed-on: https://review.couchbase.org/c/couchnode/+/234780
Tested-by: Mateusz <matt.wozakowski@couchbase.com>
Reviewed-by: Jared Casey <jared.casey@couchbase.com>